% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectrum.R
\name{FrequencyScale}
\alias{FrequencyScale}
\title{Frequency scale used to interpret the frequency bins}
\usage{
FrequencyScale(signal, window)
}
\arguments{
\item{signal}{Source signal.}

\item{window}{Window used to compute the power spectrum.}
}
\value{
A vector of frequencies \code{window$size} long.
}
\description{
The resulting frequencies are relative to the carrier frequency of the
signal, if present, otherwise to 0.
}

