% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file.R
\name{FromFile}
\alias{FromFile}
\title{Create a signal representation from a file.}
\usage{
FromFile(file.name, sample.rate, carrier.frequency = NA)
}
\arguments{
\item{file.name}{Input file name.}

\item{sample.rate}{Sample rate.}

\item{carrier.frequency}{Carrier frequency.}
}
\value{
A signal representation.
}
\description{
Create a signal representation from a file.
}

