% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signal.R
\name{FromRawVector}
\alias{FromRawVector}
\title{Create a signal representation from a raw vector of bytes.}
\usage{
FromRawVector(raw.vector, sample.rate, carrier.frequency = NA)
}
\arguments{
\item{raw.vector}{Input vector.}

\item{sample.rate}{Sample rate.}

\item{carrier.frequency}{Carrier frequency.}
}
\value{
A signal representation.
}
\description{
The raw vector must be a sequence of I/Q values (1 byte per component).
}

