% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/windows.R
\name{Hamming}
\alias{Hamming}
\title{Hamming window function}
\usage{
Hamming(size, alpha = 0.54)
}
\arguments{
\item{size}{Window width in number of samples.}

\item{alpha}{Hamming window coefficient.}
}
\description{
Hamming window function
}
\seealso{
Other FFT: \code{\link{Hanning}},
  \code{\link{Rectangular}}, \code{\link{Spectrogram}},
  \code{\link{Spectrum}}, \code{\link{Triangular}}
}

