% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/windows.R
\name{Hanning}
\alias{Hanning}
\title{Hanning window function}
\usage{
Hanning(size)
}
\arguments{
\item{size}{Window width in number of samples.}
}
\description{
Hanning window function
}
\seealso{
Other FFT: \code{\link{Hamming}},
  \code{\link{Rectangular}}, \code{\link{Spectrogram}},
  \code{\link{Spectrum}}, \code{\link{Triangular}}
}

