% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operators.R
\name{InPhase}
\alias{InPhase}
\title{In-phase component of an I/Q sample}
\usage{
InPhase(iq)
}
\arguments{
\item{iq}{I/Q complex value.}
}
\value{
In-phase component of an I/Q sample.
}
\description{
In-phase component of an I/Q sample
}

