% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iq-package.R
\docType{package}
\name{iq-package}
\alias{iq}
\alias{iq-package}
\title{I/Q file analysis toolkit}
\description{
This package offers a convenient way to parse and analyze I/Q files like
those produced by an RTL-SDR dongle.
}

