% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file.R
\name{LoadRawFile}
\alias{LoadRawFile}
\title{Load a binary file into a raw vector.}
\usage{
LoadRawFile(file.name)
}
\arguments{
\item{file.name}{Input file name.}
}
\value{
A raw vector containing the file.
}
\description{
Accordingly to the file extension, the content of the file is passed to
\code{memDecompress}.
}

