% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operators.R
\name{Magnitude}
\alias{Magnitude}
\title{Magnitude of an I/Q sample}
\usage{
Magnitude(iq)
}
\arguments{
\item{iq}{I/Q complex value.}
}
\value{
Magnitude of an I/Q sample.
}
\description{
Magnitude of an I/Q sample
}

