% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization.R
\name{NormalizeRawSamples}
\alias{NormalizeRawSamples}
\title{Normalize raw samples}
\usage{
NormalizeRawSamples(raw.vector)
}
\arguments{
\item{raw.vector}{Input}
}
\value{
A vector of normalized samples.
}
\description{
Convert the given samples from a raw vector of unisgned bytes to numerics in
a range from -1 to 1 (inclusive).
}

