% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operators.R
\name{Quadrature}
\alias{Quadrature}
\title{Quadrature component of an I/Q sample}
\usage{
Quadrature(iq)
}
\arguments{
\item{iq}{I/Q complex value.}
}
\value{
Quadrature component of an I/Q sample.
}
\description{
Quadrature component of an I/Q sample
}

