% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/windows.R
\name{Rectangular}
\alias{Rectangular}
\title{Rectangular window function}
\usage{
Rectangular(size)
}
\arguments{
\item{size}{Window width in number of samples.}
}
\description{
Rectangular window function
}
\seealso{
Other FFT: \code{\link{Hamming}}, \code{\link{Hanning}},
  \code{\link{Spectrogram}}, \code{\link{Spectrum}},
  \code{\link{Triangular}}
}

