% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signal.R
\name{Resample}
\alias{Resample}
\title{Create a new signal by resampling a portion of another}
\usage{
Resample(signal, new.sample.rate, first.sample.index = NA,
  last.sample.index = NA, from.time = NA, to.time = NA)
}
\arguments{
\item{signal}{Source signal}

\item{new.sample.rate}{New sample rate.}

\item{first.sample.index}{First sample index.}

\item{last.sample.index}{Last sample index.}

\item{from.time}{Time of the first sample.}

\item{to.time}{Time of the last sample.}
}
\value{
A resampled signal.
}
\description{
Range parameters are interpreted by \code{\link{SampleIndexRange}}.
}

