% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranges.R
\name{SampleIndexAtTime}
\alias{SampleIndexAtTime}
\title{Sample index at the given time instant.}
\usage{
SampleIndexAtTime(time, sample.rate)
}
\arguments{
\item{time}{Time instant.}

\item{sample.rate}{Sample rate of the signal.}
}
\value{
A sample index at the given time instant.
}
\description{
Sample index at the given time instant.
}

