% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranges.R
\name{SampleIndexRange}
\alias{SampleIndexRange}
\title{Extract a range of sample indexes.}
\usage{
SampleIndexRange(signal, first.sample.index, last.sample.index, from.time,
  to.time, max.samples)
}
\arguments{
\item{signal}{Source signal}

\item{first.sample.index}{Time of the first sample.}

\item{last.sample.index}{Time of the last sample.}

\item{from.time}{Time of the first sample.}

\item{to.time}{Time of the last sample.}

\item{max.samples}{Maximum number of returned sample indices.}
}
\value{
A vector containing at most \code{max.samples} sample indices from
  the given range.
}
\description{
Extract a range of sample indexes.
}

