% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranges.R
\name{SampleRange}
\alias{SampleRange}
\title{Extract a range of the signal}
\usage{
SampleRange(signal, first.sample.index = NA, last.sample.index = NA,
  from.time = NA, to.time = NA, max.samples = NA, sample.rate = NA)
}
\arguments{
\item{signal}{Source signal}

\item{first.sample.index}{First sample index.}

\item{last.sample.index}{Last sample index.}

\item{from.time}{Time of the first sample.}

\item{to.time}{Time of the last sample.}

\item{max.samples}{Maximum number of returned sample indices.}

\item{sample.rate}{New sample rate.}
}
\value{
A named list containing two items: \code{time} and \code{iq}.
}
\description{
Samples outside the signal range are assumed to be zero. Range parameters are
interpreted by \code{\link{SampleIndexRange}}.
}

