% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectrum.R
\name{Spectrogram}
\alias{Spectrogram}
\title{Compute the spectrogram of the signal at the given time range.}
\usage{
Spectrogram(signal, from.time = 0, to.time = signal$duration,
  window = Rectangular(1024), segment.overlap = 0, n.segments)
}
\arguments{
\item{signal}{Source signal.}

\item{from.time}{From time instant.}

\item{to.time}{To time instant.}

\item{window}{FFT window function.}

\item{segment.overlap}{Segment overlap ratio (0 means contiguous segments),
negative values produce spaced segments.}

\item{n.segments}{Exact number of segments (take precedence over
\code{segment.overlap} which is computed accordingly).}
}
\value{
A named list suitable to be printed with \code{image}, \code{x}
  element represents the time while \code{y} represents the frequency.
}
\description{
Compute the spectrogram of the signal at the given time range.
}
\seealso{
graphics::image

Other FFT: \code{\link{Hamming}}, \code{\link{Hanning}},
  \code{\link{Rectangular}}, \code{\link{Spectrum}},
  \code{\link{Triangular}}
}

