% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectrum.R
\name{Spectrum}
\alias{Spectrum}
\title{Compute the power spectrum of the signal at the given time}
\usage{
Spectrum(signal, at.time, window = Rectangular(1024))
}
\arguments{
\item{signal}{Source signal.}

\item{at.time}{Time instant.}

\item{window}{FFT window function. Custom windows are just named lists:
\code{'size'} the window size in number of samples; \code{'value'} a
function taking an integer in range \code{0:(size - 1)} and returning the
window value.}
}
\value{
A vector of \code{window$size} elements containing the resulting FFT
  frequency bins.
}
\description{
Compute the power spectrum of the signal at the given time
}
\seealso{
Other FFT: \code{\link{Hamming}}, \code{\link{Hanning}},
  \code{\link{Rectangular}}, \code{\link{Spectrogram}},
  \code{\link{Triangular}}
}

