% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranges.R
\name{TimeAtSampleIndex}
\alias{TimeAtSampleIndex}
\title{Time corresponding to the beginning of the given sample.}
\usage{
TimeAtSampleIndex(sample.index, sample.rate)
}
\arguments{
\item{sample.index}{Index of the sample.}

\item{sample.rate}{Sample rate of the signal.}
}
\value{
A time corresponding to the beginning of the given sample.
}
\description{
Time corresponding to the beginning of the given sample.
}

