% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file.R
\name{ToFile}
\alias{ToFile}
\title{Write a signal representation to a file.}
\usage{
ToFile(signal, file.name)
}
\arguments{
\item{signal}{Source signal.}

\item{file.name}{Output file name.}
}
\description{
The file is compressed according to its extension.
}
\seealso{
\code{\link{base::memCompress}}
}

