% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/windows.R
\name{Triangular}
\alias{Triangular}
\title{Triangular window function}
\usage{
Triangular(size)
}
\arguments{
\item{size}{Window width in number of samples.}
}
\description{
Triangular window function
}
\seealso{
Other FFT: \code{\link{Hamming}}, \code{\link{Hanning}},
  \code{\link{Rectangular}}, \code{\link{Spectrogram}},
  \code{\link{Spectrum}}
}

