% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranges.R
\name{SampleIndexRange}
\alias{SampleIndexRange}
\title{Extract a range of sample indexes.}
\usage{
SampleIndexRange(signal, first.sample.index, last.sample.index, from.time,
  to.time, max.samples)
}
\arguments{
\item{signal}{Source signal}

\item{first.sample.index}{First sample index.}

\item{last.sample.index}{Last sample index.}

\item{from.time}{Time of the first sample.}

\item{to.time}{Time of the last sample.}

\item{max.samples}{Maximum number of returned sample indices.}
}
\value{
A vector containing at most \code{max.samples} sample indices from
  the given range.
}
\description{
Sample indices take precedence over times. When boudaries are omitted
(\code{NA}) then the boundaries of the signal are used.
}
\details{
If \code{max.samples} is \code{NA} all the sample indices in the range are
returned.
}

