% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranges.R
\name{SampleRange}
\alias{SampleRange}
\title{Extract a range of the signal.}
\usage{
SampleRange(signal, first.sample.index = 1,
  last.sample.index = signal$n.samples, from.time = 0,
  to.time = signal$duration, max.samples = Inf)
}
\arguments{
\item{signal}{Source signal}

\item{first.sample.index}{Time of the first sample.}

\item{last.sample.index}{Time of the last sample.}

\item{from.time}{Time of the first sample.}

\item{to.time}{Time of the last sample.}

\item{max.samples}{Maximum number of returned sample indices.}
}
\value{
A named list containing two items: \code{time} and \code{iq}.
}
\description{
Samples outside the signal range are assumed to be zero.
}
\details{
Sample indices take precedence over times.
}

